refactor: moved badge color utilities from storage to Badge component
